Convert localization

+++ src/api/java/thaumcraft/api/crafting/InfusionRecipe.java
@@ -94,7 +94,7 @@ public class InfusionRecipe {
 			t1 = true;
 			int od = OreDictionary.getOreID(stack0);
 			if (od != -1) {
-				ItemStack[] ores = OreDictionary.getOres(od).toArray(new ItemStack[] {});
+				ItemStack[] ores = OreDictionary.getOres((Integer)od).toArray(new ItemStack[] {});
 				if (ThaumcraftApiHelper.containsMatch(false, new ItemStack[] { stack1 }, ores)) {
 					return true;
 				}
